+++ src/ace_time/common/compat.h
@@ -60,7 +60,7 @@
     #define SERIAL_PORT_MONITOR Serial
   #endif
 
-#elif defined(UNIX_HOST_DUINO)
+#elif defined(EPOXY_DUINO)
   #include <pgmspace.h>
 
 #elif defined(ARDUINO_ARCH_STM32)